% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class_label.R
\name{class_label}
\alias{class_label}
\title{A MVDA Function}
\usage{
class_label(CM)
}
\arguments{
\item{CM}{is the confusion matrix}
}
\value{
a vector containing a label for each cluster
}
\description{
This function assign labels to each cluster. the label is assigned to the majority. That is, for each cluster identifies the predominance of objects of a given class. The cluster will be assigned the same label of the predominant class.
}

