% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster_entropia.R
\name{cluster_entropia}
\alias{cluster_entropia}
\title{A MVDA Function}
\usage{
cluster_entropia(vector.clust)
}
\arguments{
\item{cluster.vector}{a numeric vector of clustering results}
}
\value{
the entropy in the cluster
}
\description{
This function evaluate the entropy of a cluster
}
\keyword{clustering;}
\keyword{entropy}
\keyword{multi-view}

