% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster_mutual_information.R
\name{cluster_mutual_information}
\alias{cluster_mutual_information}
\title{A MVDA Function}
\usage{
cluster_mutual_information(vc1, vc2, elem)
}
\arguments{
\item{vc1}{a numeric vector of clustering results}

\item{vc2}{a numeric vector of clustering results}

\item{elem}{the matrix representing your dataset}
}
\value{
the mutual information between clustering
}
\description{
This function evaluate the mutual information between two clustering results
}
\keyword{clustering;}
\keyword{entropy}
\keyword{multi-view}

