% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clustering_summary.R
\name{clustering_summary}
\alias{clustering_summary}
\title{A MVDA Function}
\usage{
clustering_summary(DB, cluster)
}
\arguments{
\item{DB}{your matrix dataset}

\item{cluster}{is a numeric vector of clustering results}
}
\value{
a vector of evaluation indexes
}
\description{
This function evaluate the summary intdex for a clustering results
}
\keyword{clustering}
\keyword{evaluation}

