% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CMsup_error.R
\name{CMsup_error}
\alias{CMsup_error}
\title{A MVDA Function}
\usage{
CMsup_error(CM_sup, nPat)
}
\arguments{
\item{CM_sup}{confusion matrix}

\item{nPat}{number of patients}
}
\value{
an error value
}
\description{
This function evaluate the error in a confusion matrix
}
\keyword{clustering;}
\keyword{confusion}
\keyword{matrix}
\keyword{multi-view}

