% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{A MVDA Function}
\usage{
confusion_matrix(etichette, clustering, pazienti_geni, nCluster)
}
\arguments{
\item{etichette}{is a vector of patient labels}

\item{clustering}{is a vector of clustering results}

\item{pazienti_geni}{is your matrix dataset}

\item{nCluster}{is the number of obtained clusters}
}
\value{
the confusion matrix
}
\description{
This function construct confusion matrix between patient classes and the obtained clustering
}
\keyword{clustering;}
\keyword{confusion}
\keyword{matrix}
\keyword{multi-view}

