% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/consensus_matrix.R
\name{consensus_matrix}
\alias{consensus_matrix}
\title{A MVDA Function}
\usage{
consensus_matrix(cluster_list)
}
\arguments{
\item{cluster_list}{is list of vector of clustering results}
}
\value{
a consensus matrix
}
\description{
This function consensus matrix from a list of clustering results. It evaluetes for each couple of patients
how many times they have been clustered together.
}
\keyword{clustering;}
\keyword{confusion}
\keyword{matrix}
\keyword{multi-view}

