% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/matrix_factorization_by_NNSVD.R
\name{error}
\alias{error}
\title{A MVDA Function}
\usage{
error(X, PH, l, n)
}
\arguments{
\item{X}{is the matrix resulting from the concatenation of the membership matrix for each view}

\item{PH}{is the approximation of X}

\item{l}{is the number of rows of X}

\item{n}{is the number of columns of X}
}
\value{
the approximation error
}
\description{
This function calculates the approximation error at each iteration as the difference between the matrix X and its approximation PH
}

