% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/matrix_factorization_by_NNSVD.R
\name{find_k}
\alias{find_k}
\title{A MVDA Function}
\usage{
find_k(kmin, kmax, X, eps)
}
\arguments{
\item{kmin}{is the value of k from wich it start to search}

\item{kmax}{is the maximum value of k}

\item{X}{is the matrix resulting from the concatenation of the membership matrix for each view}

\item{eps}{is the precision parameter}
}
\value{
the best factorization and the optimum value of k
}
\description{
This function calculates the best value of k through a series of executions (from kmin to kmax) of the algorithm of factorization in each step calculates the entropy of the matrix P of Responsibilities and chooses the partition with greater entropy value
}

