% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/matrix_factorization_by_NNSVD.R
\name{find_T}
\alias{find_T}
\title{A MVDA Function}
\usage{
find_T(P, k, nView, V1.lc1, V2.lc2, V3.lc3, V4.lc4, V5.lc5)
}
\arguments{
\item{P}{is the projection matrix of the approximation PH}

\item{k}{is the number of cluster}

\item{nView}{is the number of views in the experiments}

\item{V1.lc1}{is the number of cluster in the first view}

\item{V2.lc2}{is the number of cluster in the second view}

\item{V3.lc3}{is the number of cluster in the third view and so on...}
}
\value{
the approximation error
}
\description{
This function calculates the matrix T indicating the contributing ot each view to each meta-cluster
}

