% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findCenter.R
\name{findCenter}
\alias{findCenter}
\title{A MVDA Function}
\usage{
findCenter(DB, clust_vector)
}
\arguments{
\item{DB}{your matrix dataset}

\item{cluster.vector}{a numeric vector of clustering results}
}
\value{
a matrix of prototypes
}
\description{
This function select cluster prototypes. It select the prototype as the more correlated with the other elements in the cluster
}
\keyword{clustering;}
\keyword{correlation}
\keyword{multi-view}
\keyword{prototype;}

