% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fisher.R
\name{fisher_test}
\alias{fisher_test}
\title{A MVDA Function}
\usage{
fisher_test(clust, cl.cl, CM_sup, info)
}
\arguments{
\item{clust}{is the vector of clustering assignment}

\item{CM_sup}{is the confusion matrix}

\item{info}{is the factor of patient classes}

\item{cl.cla}{is the vector of labels for each cluster}
}
\value{
a matrix containing a pvalue for each cluster for each class label.
}
\description{
run the fisher's exact test for each cluster in order to assign statistical significance to the label that has been assigned.
}

