% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/general_late_integration.R
\name{general_late_integration}
\alias{general_late_integration}
\title{A MVDA Function}
\usage{
general_late_integration(A, k, alfa, eps, pazienti, info)
}
\arguments{
\item{A}{is the matrix resulting from the concatenation of the membership matrix for each view}

\item{k}{the desidered number of cluster}

\item{eps}{is the precision parameter}
}
\value{
a list containing the general linear integration results results, the clustering assignment, the confusion matrix, the error in confusion matrix, the matrix of clusters centroids, the label for each cluster and the matrix with fisher test results for each cluster
}
\description{
This function implements the general linear integration process
}

