% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gli_function.R
\name{GLI}
\alias{GLI}
\title{A MVDA Function}
\usage{
GLI(A, k, alfa, eps)
}
\arguments{
\item{A}{is a matrix in R^n*r (where n is the number of elements and r is the number of clusters) of m membership matrix stacked orizontally}

\item{k}{the desidered number of cluster}

\item{alfa}{that is a positive constant}

\item{eps}{is the constant parameter for convergence}
}
\value{
a list containing two matrix: - A clustering pattern matrix B (in R^ n*k) - A matrix P (in R^ k*r) of m mapping matrices stacked orizontally
}
\description{
Function that perform che multiple view late integration clustering
}

