% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kmeans_preprocessing.R
\name{kmeans_preprocessing}
\alias{kmeans_preprocessing}
\title{A MVDA Function}
\usage{
kmeans_preprocessing(DB = NULL, nCenters = NULL, iter.max = 100,
  nstart = 10, method = "pearson")
}
\arguments{
\item{DB}{is your matrix dataset}

\item{nCenters}{is the desidered number of cluster}

\item{iter.max}{The maximum number of iterations allowed to Kmeans}

\item{nstart}{If nCenter is a number, how many random sets should be chosen?}

\item{method}{is the method by wich distance is evaluated. Default is pearson.}
}
\value{
a list containing three field: pamk.res is the pamk results. clustering is the vector with clustering assignment. center is the matrix with center prototypes.
}
\description{
This function execute kmeans clustering on single view patient prorotypes. It require library amap.
}
\keyword{kmeans-clustering;}
\keyword{ward}

