% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kmeans_sv.R
\name{kmeans_sv}
\alias{kmeans_sv}
\title{A MVDA Function}
\usage{
kmeans_sv(nCenters, prototype, method = "pearson",
  corr.use = "pairwise.complete.obs", iter.max = 10000, nstart = 10)
}
\arguments{
\item{nCenters}{is the number of cluster we want to obtain}

\item{prototype}{is the matrix of prototype we want to cluster}

\item{method}{is the method by wich distance is evaluated. Default is pearson.}

\item{corr.use}{is an optional character string giving a method for computing covariances in the presence of missing values. This must be (an abbreviation of) one of the strings "everything", "all.obs", "complete.obs", "na.or.complete", or "pairwise.complete.obs". Default value="pairwise.complete.obs"}

\item{iter.max}{The maximum number of iterations allowed to Kmeans}

\item{nstart}{If nCenter is a number, how many random sets should be chosen?}
}
\value{
a list containing three field: pamk.res is the pamk results. clustering is the vector with clustering assignment. center is the matrix with center prototypes.
}
\description{
This function execute kmeans clustering on single view patient prorotypes. It require library amap.
}
\keyword{kmeans-clustering;}
\keyword{ward}

