% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/matrix_factorization_by_NNSVD.R
\name{matrix_factorization}
\alias{matrix_factorization}
\title{A MVDA Function}
\usage{
matrix_factorization(X, k1, eps, iter_max)
}
\arguments{
\item{X}{is the matrix resulting from the concatenation of the membership matrix for each view}

\item{k1}{the desidered number of cluster}

\item{eps}{is the precision parameter}

\item{iter_mat}{the maximum allowed iteration}
}
\value{
a list containing the two matri P and H
}
\description{
This function implements the matrix factorization algorithm
}

