% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mean_correlation.R
\name{mean_correlation}
\alias{mean_correlation}
\title{A MVDA Function}
\usage{
mean_correlation(DB, clust_vector, center)
}
\arguments{
\item{DB}{your matrix dataset}

\item{center}{a matrix with cluster prototype}

\item{cluster.vector}{a numeric vector of clustering results}
}
\value{
the vector of mean correlation between object and prototypes
}
\description{
This function find correlation between object in each cluster and its prototype
}
\keyword{clustering;}
\keyword{correlation}
\keyword{feature-relevance;}
\keyword{multi-view}

