% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/matrix_factorization.R
\name{MF}
\alias{MF}
\title{A MVDA Function}
\usage{
MF(k, A, eps, iter_max, nView, V1.lc1, V1.lc2, V1.lc3, V1.lc4, V1.lc5, pazienti,
  info)
}
\arguments{
\item{k}{the desidered number of cluster}

\item{A}{is the matrix resulting from the concatenation of the membership matrix for each view}

\item{eps}{is the precision parameter}

\item{nView}{is the number of views}

\item{pazienti}{is your matrix dataset}

\item{info}{is the factor of class labels}

\item{iter_mat}{the maximum allowed iteration}
}
\value{
a list containing the matrix factorization results, the clustering assignment, the confusion matrix, the error in confusion matrix, the matrix T of views influence, the matrix of clusters centroids, the labels for each cluster and the matrix of fisher test for each cluster
}
\description{
This function implements the matrix factorization integration process
}

