% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MV_clustering_stability.R
\name{mvclst}
\alias{mvclst}
\title{A MVDA Function}
\usage{
mvclst(method = "matrix_factorization", center, class_cluster, X)
}
\arguments{
\item{method}{is the method you can use for the integration. Values can be matrix_factorization (default) of gli.}

\item{center}{is the matrix of centroid}

\item{class_cluster}{is the vector of class labels assigned to each cluster}

\item{X}{is the staked matrix of membeship matrix of each view.}
}
\value{
a list containing the list of clustering obtained, the NxN matrix of NMI betwee each couple of clustering, the mean and the standard deviation of this matrix.
}
\description{
This function evaluate clustering stability by performing multi-view integration perturbing the dataset via leave-one-out and comparing
NMI between different clustering results.
}

