% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/OrderByCorrelation.R
\name{OrderByCorrelation}
\alias{OrderByCorrelation}
\title{A MVDA Function}
\usage{
OrderByCorrelation(n_cluster, cluster.pv, DB, cluster.vector)
}
\arguments{
\item{DB}{your matrix dataset}

\item{cluster.vector}{a numeric vector of clustering results}
}
\value{
a list with two object prototipi_geni is the list of feature ordered by correlation and pazienti that is the list of patients ordered by correlations
}
\description{
This function allows you to rank feature relevance for each multi-view cluster by correlation.
It works by removing a feature at a time and calculating correlation between patients in the cluster.
The more the correlation decrease, the more the feature is relevant for the clusters.
It also order patients by correlation.
}
\keyword{clustering;}
\keyword{correlation}
\keyword{feature-relevance;}
\keyword{multi-view}

