% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pamk_prep.R
\name{pamk_preprocessing}
\alias{pamk_preprocessing}
\title{A MVDA Function}
\usage{
pamk_preprocessing(DB = NULL, nCenters = NULL, initial_medoids = NULL)
}
\arguments{
\item{DB}{is your matrix dataset}

\item{nCenters}{is the desidered number of cluster}

\item{initial_medoids}{index of the desidered initial medoids you want to use to initialize the algorithm}
}
\value{
a list containing three field: pamk.res is the pamk results. clustering is the vector with clustering assignment. center is the matrix with center prototypes.
}
\description{
This function execute partitioning around medoids algoritm using function from fpc package.
}
\keyword{pamk;}
\keyword{ward}

