% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pamk_sv.R
\name{pamk_sv}
\alias{pamk_sv}
\title{A MVDA Function}
\usage{
pamk_sv(nCenters, prototype, dist.method = "pearson",
  corr.use = "pairwise.complete.obs")
}
\arguments{
\item{nCenters}{is the number of cluster we want to obtain}

\item{prototype}{is the matrix of prototype we want to cluster}

\item{dist.method}{is the method by wich distance is evaluated. Default is pearson.}

\item{corr.use}{is an optional character string giving a method for computing covariances in the presence of missing values. This must be (an abbreviation of) one of the strings "everything", "all.obs", "complete.obs", "na.or.complete", or "pairwise.complete.obs". Default value="pairwise.complete.obs"}
}
\value{
a list containing three field: pamk.res is the pamk results. clustering is the vector with clustering assignment. center is the matrix with center prototypes.
}
\description{
This function execute partitioning around medoids clustering with on single view patient prorotypes. It require library fpc.
}
\keyword{pamk-clustering;}
\keyword{ward}

