% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotCM.R
\name{plotCM}
\alias{plotCM}
\title{A MVDA Function}
\usage{
plotCM(CM, xlab = "Clusters", ylab = "Pam50",
  title = "Patients Classes Confusion Matrix", toSave = F,
  cartella = "./MV_clust/Kmeans/")
}
\arguments{
\item{CM}{is the confusion matrix}

\item{toSave}{if you want to save or not the image in a file. Default is FALSE}

\item{cartela}{is the path where the file will be saved}
}
\value{
a ggplot object
}
\description{
This function plot a confusion matrix
}
\keyword{clustering;}
\keyword{confusion}
\keyword{matrix}
\keyword{multi-view}

