% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/printStatistics.R
\name{printStatistics}
\alias{printStatistics}
\title{A MVDA Function}
\usage{
printStatistics(DB, clust_vector, center, cartella)
}
\arguments{
\item{DB}{your matrix dataset}

\item{center}{a matrix with cluster prototype}

\item{cartella}{is the path where the function will save the statistics files}

\item{cluster.vector}{a numeric vector of clustering results}
}
\description{
This function print statistics about clustering results
}
\keyword{clustering;}
\keyword{correlation}
\keyword{feature-relevance;}
\keyword{multi-view}

