% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pvclust_prep.R
\name{pvclust_prep}
\alias{pvclust_prep}
\title{A MVDA Function}
\usage{
pvclust_prep(DB, hclust.method = "ward", nboot = 100,
  dist.method = "correlation", alpha = 0.95, show.print = F, r = 1,
  printLocation = ".", km_center = F, nCenters)
}
\arguments{
\item{DB}{is your matrix dataset}

\item{hclust.method}{is the the agglomeration method to be used for the hclust function. Default value is "ward".}

\item{nboot}{is the number of bootstrap replications. The default is 100.}

\item{dist.method}{is the distance measure to be used.}

\item{r}{is numeric vector which specifies the relative sample sizes of bootstrap replications. The default is 1.}

\item{nCenters}{is the desidered number of cluster}

\item{nCenters}{is the desidered number of clusters}
}
\value{
a list containing three field: pvclust.res is the pvclust clustering results. clustering is the vector with clustering assignment. center is the matrix with center prototypes.
}
\description{
This function calculate clustering of single view patient prototypes by using pvclust packages.
}
\keyword{correlation;}
\keyword{hierarchical-clustering;}
\keyword{p-value}
\keyword{pvclust;}
\keyword{ward;}

