% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pvclust_sv.R
\name{pvclust_sv}
\alias{pvclust_sv}
\title{A MVDA Function}
\usage{
pvclust_sv(prototype, hls_method = "ward", nboot = 100, t = 0.95,
  print = F, r = 1, print_folder = ".", dist.method = "correlation")
}
\arguments{
\item{prototype}{is the matrix of prototype we want to cluster}

\item{hls_method}{is the the agglomeration method to be used for the hclust function. Default value is "ward".}

\item{nboot}{is the number of bootstrap replications. The default is 100.}

\item{r}{is numeric vector which specifies the relative sample sizes of bootstrap replications. The default is 1.}

\item{dist.method}{is the distance measure to be used.}
}
\value{
a list containing three field: pvclust.res is the pvclust clustering results. clustering is the vector with clustering assignment. center is the matrix with center prototypes.
}
\description{
This function calculate clustering of single view patient prototypes by using pvclust packages.
}
\keyword{correlation;}
\keyword{hierarchical-clustering;}
\keyword{p-value}
\keyword{pvclust;}
\keyword{ward;}

