% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rf_ranking.R
\name{rf_ranking}
\alias{rf_ranking}
\title{A MVDA Function}
\usage{
rf_ranking(prototype, info, MTRY = c(1, 2, 3), NTREE = c(100, 200, 500),
  NPERM = c(1, 2, 3), NODESIZE = c(1, 5))
}
\arguments{
\item{prototype}{is the matrix of prototype, patients are on columns and featuers are on rows}

\item{info}{is the factor of class label}

\item{MTRY}{vector of possible values for the random forest's mtry parameters}

\item{NTREE}{vector of possible values for the random forest's ntree parameters}

\item{NPERM}{vector of possible values for the random forest's nperm parameters}

\item{NODESIZE}{vector of possible values for the random forest's nodesize parameters}
}
\value{
A matrix of importance measure, one row for each predictor variable. The column(s) are different importance measures.
}
\description{
This function determines a ranking of predictors by using random foresto from the package randomForest.
}
\keyword{clustering;}
\keyword{feature}
\keyword{multi-view}
\keyword{randomForest;}
\keyword{ranking;}

