% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sda_ranking.R
\name{sda_ranking}
\alias{sda_ranking}
\title{A MVDA Function}
\usage{
sda_ranking(prototype = NULL, info = NULL, fdr = FALSE,
  ranking.score = "avg", lambda = 0.5)
}
\arguments{
\item{prototype}{is the matrix of prototype}

\item{info}{is the factor of class label}

\item{fdr}{compute FDR values and HC scores for each feature. default = FALSE}

\item{ranking.score}{how to compute the summary score for each variable from the CAT scores of all classes. default = "avg"}

\item{lambda}{Shrinkage intensity for the correlation matrix. If not specified it is estimated from the data. lambda=0 implies no shrinkage and lambda=1 complete shrinkage. default = 0.5}
}
\value{
return a matrix with the following columns: idx    original feature number  score  sum of the squared CAT scores across groups - this determines the overall ranking of a feature cat for each group and feature the cat score of the centroid versus the pooled mean
}
\description{
This function determines a ranking of predictors by computing CAT scores (correlation-adjusted t-scores) between the group centroids and the pooled mean by using fucntion of the package sda.
}
\keyword{cat-score;}
\keyword{clustering;}
\keyword{feature}
\keyword{multi-view}
\keyword{ranking;}

