% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spectral_preprocessing.R
\name{spectral_preprocessing}
\alias{spectral_preprocessing}
\title{A MVDA Function}
\usage{
spectral_preprocessing(DB = NULL, nCenters = NULL, kernel_f = "rbfdot")
}
\arguments{
\item{DB}{is your matrix dataset}

\item{nCenters}{is the desidered number of cluster}

\item{kernel_f}{is the kernel function used in computing the affinity matrix. Default value: rbfdot}
}
\value{
a list containing three field: specc.res is the spectral clustering results. clustering is the vector with clustering assignment. center is the matrix with center prototypes.
}
\description{
This function execute spectral clustering  on single view patient prorotypes
}
\keyword{spectral-clustering;}

