% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spectral_sv.R
\name{spectral_sv}
\alias{spectral_sv}
\title{A MVDA Function}
\usage{
spectral_sv(nCenters, prototype, kernel_function = "rbfdot")
}
\arguments{
\item{nCenters}{is the number of cluster we want to obtain}

\item{prototype}{is the matrix of prototype we want to cluster}

\item{kernel_function}{is the kernel function used in computing the affinity matrix. Default value: rbfdot}
}
\value{
a list containing three field: specc.res is the spectral clustering results. clustering is the vector with clustering assignment. center is the matrix with center prototypes.
}
\description{
This function execute spectral clustering  on single view patient prorotypes
}
\keyword{spectral-clustering;}

