% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/supervised_matrix.R
\name{supervised_matrix}
\alias{supervised_matrix}
\title{A MVDA Function}
\usage{
supervised_matrix(etichette, nRow, nCol, row_id, col_id)
}
\arguments{
\item{etichette}{is the vector with class labels}

\item{nRow}{is the number of rows of the matrix (number of patients)}

\item{nCol}{is the number of columns of the matrix (number of classes)}

\item{row_id}{are the rows names}

\item{col_id}{are the cols names}
}
\value{
a membership matrix
}
\description{
This function create the membership matrix of patient and classes
}
\keyword{clustering;}
\keyword{confusion}
\keyword{matrix}
\keyword{multi-view}

