% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ward_prep.R
\name{ward_hls_prep}
\alias{ward_hls_prep}
\title{A MVDA Function}
\usage{
ward_hls_prep(DB = NULL, nCenters = NULL)
}
\arguments{
\item{DB}{is your matrix dataset}

\item{nCenters}{is the desidered number of cluster}
}
\value{
a list containing three field: hls.res is the hclust results. clustering is the vector with clustering assignment. center is the matrix with center prototypes.
}
\description{
This function execute hierarchical clustering with ward method
}
\keyword{hierarchical-clustering;}
\keyword{ward}

