% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ward_sv.R
\name{ward_sv}
\alias{ward_sv}
\title{A MVDA Function}
\usage{
ward_sv(nCenters, prototype, corr_method = "pearson",
  use = "pairwise.complete.obs", hls_method = "ward")
}
\arguments{
\item{nCenters}{is the number of cluster we want to obtain}

\item{prototype}{is the matrix of prototype we want to cluster}

\item{corr_method}{is the method by wich distance is evaluated. Default is pearson.}

\item{use}{is an optional character string giving a method for computing covariances in the presence of missing values. This must be (an abbreviation of) one of the strings "everything", "all.obs", "complete.obs", "na.or.complete", or "pairwise.complete.obs". Default value="pairwise.complete.obs"}

\item{hls_method}{is the the agglomeration method to be used for the hclust function. Default value is "ward".}
}
\value{
a list containing three field: hls.res is the hclust results. clustering is the vector with clustering assignment. center is the matrix with center prototypes.
}
\description{
This function execute hierarchical clustering with ward method on single view patient prorotypes
}
\keyword{hierarchical-clustering;}
\keyword{ward}

